+++ test/plausible_web/controllers/api/external_stats_controller/aggregate_test.exs
@@ -588,28 +588,6 @@ defmodule PlausibleWeb.Api.ExternalStatsController.AggregateTest do
                "visitors" => %{"value" => 1, "change" => 100}
              }
     end
-
-    test "events metric with imported data is disallowed", %{
-      conn: conn,
-      site: site,
-      site_import: site_import
-    } do
-      populate_stats(site, site_import.id, [
-        build(:imported_visitors, date: ~D[2023-01-01])
-      ])
-
-      conn =
-        get(conn, "/api/v1/stats/aggregate", %{
-          "site_id" => site.domain,
-          "period" => "day",
-          "date" => "2023-01-02",
-          "metrics" => "events",
-          "with_imported" => "true"
-        })
-
-      assert %{"error" => msg} = json_response(conn, 400)
-      assert msg == "Metric `events` cannot be queried with imported data"
-    end
   end
 
   describe "filters" do