+++ CHANGELOG.md
@@ -41,6 +41,10 @@ All notable changes to this project will be documented in this file.
 - Add restrictive `robots.txt` for self-hosted plausible/analytics#3905
 - Add Yesterday as an time range option in the dashboard
 - Add dmg extension to the list of default tracked file downloads
+- Add support for importing Google Analytics 4 data
+- Import custom events from Google Analytics 4
+- Ability to filter Search Console keywords by page, country and device plausible/analytics#4077
+- Add `DATA_DIR` env var for exports/imports plausible/analytics#4100
 
 ### Removed
 - Removed the nested custom event property breakdown UI when filtering by a goal in Goal Conversions